% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{addLipidProperties}
\alias{addLipidProperties}
\title{Add Lipid properties}
\usage{
addLipidProperties(
  data,
  lipidClassOrder = c("PC", "PC O-", "LPC", "PE", "PE O-", "PE P-", "LPE", "PS",
    "PS O-", "PI", "PI O-", "PG", "PG O-", "LPG", "PA", "PA O-", "LPA", "CL", "MLCL",
    "Cer", "SM", "HexCer", "SGalCer", "GM3", "Sulf", "diHexCer", "Hex2Cer", "For", "IPC",
    "MIPC", "M(IP)2C", "Chol", "Desm", "Erg", "CE", "EE", "DAG", "TAG", "PIP", "PIP2",
    "PIP3", "GM1Cer", "GD1Cer", "MAG", "Epi", "PGP", "WE", "FA")
)
}
\arguments{
\item{data}{tibble. rawData}

\item{lipidClassOrder}{character vector, order of lipid classes}
}
\value{
tibble. enhanced data
}
\description{
Based on lipid column
}
