% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collect_meta_data}
\alias{collect_meta_data}
\title{Collect metadata from database}
\usage{
collect_meta_data(con)
}
\arguments{
\item{con}{database conncetion object.
Create one yourself with e.g.
\code{DBI::dbConnect(RPostgres::Postgres(), ...)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "<database/exampleDatabase.db>")}}
}
\value{
a tibble with the meta data
}
\description{
Collect metadata from database
}
