% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{collectMetaData}
\alias{collectMetaData}
\title{Collect metadata from database}
\usage{
collectMetaData(con)
}
\arguments{
\item{con}{database conncetion object.
Create one yourself with e.g.
\code{DBI::dbConnect(RPostgres::Postgres(), ...)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "<data/exampleDatabase.db>")}}
}
\value{
a tibble with the meta data
}
\description{
Collect metadata from database
}
