% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{collectRawData}
\alias{collectRawData}
\title{Collect raw data from database}
\usage{
collectRawData(id = 1, con)
}
\arguments{
\item{id}{numeric. ID of dataset to collect}

\item{con}{A database connection}
}
\value{
tibble. Raw data
}
\description{
Collect raw data from database
}
