% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_plotData}
\alias{create_plotData}
\title{Create the data ready for plotting}
\usage{
create_plotData(df, input)
}
\arguments{
\item{df}{Input tibble}

\item{input}{a list with the aesthetic mappings.
This list can contain
strings with the following names,
non existent ones will be ignored:
\itemize{
\item \code{tecRep_average}
\item \code{aes_x}
\item \code{aes_color}
\item \code{aes_facet1}
\item \code{aes_facet2}
}}
}
\value{
neat data
}
\description{
Based on the aesthetic mappings.
}
