% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create a Heatmap from data}
\usage{
createHeatmap(plotData, input)
}
\arguments{
\item{plotData}{tibble. typically meanPlotData()}

\item{input}{:: list. Uses:
\itemize{
\item standardizationFeatures NULL | character vector.
\item aesX :: string.
\item aesColor :: string.
\item aesFacetCol :: string.
\item aesFacetRow :: string.
\item heatLabSize :: numeric. Size of axis labels
\item heatColor :: string. Color palette
One of c("viridis", "magma", "plasma", "inferno", "cividis")
\item heatLogScale :: logical.
}}
}
\value{
ggplot. Heatmap
}
\description{
Create a Heatmap from data
}
