% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{createMainPlot}
\alias{createMainPlot}
\title{Title}
\usage{
createMainPlot(
  plotData,
  meanPlotData,
  pairwiseComparisons,
  rangeX = NULL,
  rangeY = NULL,
  input
)
}
\arguments{
\item{plotData}{tibble. Data for the plot, pass it from reactive plotData()}

\item{meanPlotData}{tibble. Data of means, pass from reactive meanPlotData()}

\item{pairwiseComparisons}{tibble. Pairwise t-tests from pairwiseComparisons()}

\item{rangeX}{numeric vector | NULL. vector with min and max X}

\item{rangeY}{numeric vector | NULL.  vector with min and max Y}

\item{input}{list. Input list from shiny ui. Uses
- aesX :: string.
- aesColor :: string.
- aesFacetCol :: string.
- aesFacetRow :: string.
- mainPlotAdditionalOptions :: character vector. Options:
list("points", "bars", "mean", "values", "ind_values", "log", 
"N", "label", "swap", "free_y", "signif")
- errorbarType :: string. "None" | "SD" | "SEM" | "CI"
- summariseTechnicalReplicates :: boolean.
- standardizationFeatures :: character vector | NULL.
- standardizeWithinTechnicalReplicate :: boolean.}
}
\value{
:: ggplot object
}
\description{
Title
}
