% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{createPcaPrep}
\alias{createPcaPrep}
\title{Create prepped pca recipe}
\usage{
createPcaPrep(wideData, input)
}
\arguments{
\item{wideData}{:: tibble.}

\item{input}{:: list. Uses:
\itemize{
\item summariseTechnicalReplicates :: logical
\item aesColor
\item pcaScale
\item pcaCenter
\item pcaNumberPrincipalComponents
}}
}
\value{
:: prepped recipe
}
\description{
Create prepped pca recipe
}
