% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{createPlotData}
\alias{createPlotData}
\title{Create the data ready for plotting}
\usage{
createPlotData(data, input)
}
\arguments{
\item{data}{:: tibble}

\item{input}{:: list. Input list from shiny ui. Uses
\itemize{
\item summariseTechnicalReplicates :: boolean
\item aesX :: string
\item aesColor :: string
\item aesFacetCol :: string | ""
\item aesFacetRow :: string | ""
}}
}
\value{
:: tibble. neat data
}
\description{
Based on the aesthetic mappings.
}
