% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createQueryForID}
\alias{createQueryForID}
\title{Generate SQL Query for the selected dataset}
\usage{
createQueryForID(dataset_ID)
}
\arguments{
\item{dataset_ID}{numeric}
}
\value{
SQL Query as a string
}
\description{
Generate SQL Query for the selected dataset
}
