% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{createWideData}
\alias{createWideData}
\title{Create wide data}
\usage{
createWideData(plotData, input)
}
\arguments{
\item{plotData}{:: tibble.}

\item{input}{:: list. Uses:
\itemize{
\item aesX :: String
}}
}
\value{
:: tibble.
}
\description{
Create data in wide format for dimensionality reduction workflows
with tidymodels.
}
