% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaultInput}
\alias{defaultInput}
\title{Generate list of default options}
\usage{
defaultInput()
}
\value{
:: list. A list of default options to be assigned to \code{input}
}
\description{
When runninng the shiny app, create the global variable
\code{defaultInput <- defaultInput()}. When running
functions from ShinyLipids outside of the Shiny app,
generate the input list normally created by the web interface
with \code{input <- defaultInput()}.
}
