% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_pairwise_comparisons}
\alias{do_pairwise_comparisons}
\title{Test all possible pairwise t-tests}
\usage{
do_pairwise_comparisons(df, x_axis)
}
\arguments{
\item{df}{the full dataset}

\item{x_axis}{the feature mapped to the x-axis}
}
\value{
A tidy representation of the significance
tests build with \code{broom::tidy} combinde
into one tibble.
}
\description{
Because people like p-values and little stars.
}
