% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_handler_factories.R
\name{downloadHandlerFactoryCSV}
\alias{downloadHandlerFactoryCSV}
\title{Create Download Handler for CSV files}
\usage{
downloadHandlerFactoryCSV(metaData, dataset, specifier, id)
}
\arguments{
\item{metaData}{tibble. Meta data.}

\item{dataset}{tibble. Data to save.}

\item{specifier}{string. Postfix for the file name.}

\item{id}{integer. Id of the data set.}
}
\value{
download handler for the shiny server function
}
\description{
Create Download Handler for CSV files
}
