% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_handler_factories.R
\name{downloadHandlerFactoryPDF}
\alias{downloadHandlerFactoryPDF}
\title{Create Download Handler for PDF files}
\usage{
downloadHandlerFactoryPDF(metaData, plot, specifier, width, height, id)
}
\arguments{
\item{metaData}{tibble. Meta data.}

\item{plot}{ggplot. Plot to save.}

\item{specifier}{string. Postfix for the file name.}

\item{width}{numeric. Width in inches.}

\item{height}{numeric. Height in inches.}

\item{id}{integer. Id of the data set.}
}
\value{
download handler for the shiny server function
}
\description{
Create Download Handler for PDF files
}
