% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(df, input)
}
\arguments{
\item{df}{Data as a tibble}

\item{input}{a list.
This list can contain
strings with the following names,
non existent ones will be ignored:
\itemize{
\item \code{filter_cat}
\item \code{filter_class}
\item \code{filter_func}
\item \code{filter_length}
\item \code{filter_db}
\item \code{filter_oh}
\item \code{sample_select}
\item \code{sample_remove}
\item \code{tecRep_remove}
}}
}
\value{
Lovely filtered data
}
\description{
Filter raw data
}
