% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filterIfNotNull}
\alias{filterIfNotNull}
\title{Filter if not null}
\usage{
filterIfNotNull(data, var, condition)
}
\arguments{
\item{data}{:: tibble. data}

\item{var}{:: character vector | NULL}

\item{condition}{:: an expression to evaluate in
the context of the data}
}
\value{
tibble. filtered data
}
\description{
Filter data based on condition
only if an input var is not NULL
}
