% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(
  rawData,
  categoryToSelect = NULL,
  lipidClassToSelect = NULL,
  functionalCategoryToSelect = NULL,
  filterLengthRange = NULL,
  filterDoubleBondsRange = NULL,
  filterOhRange = NULL,
  samplesToSelect = NULL,
  samplesToRemove = NULL,
  replicatesToSelect = NULL,
  replicatesToRemove = NULL,
  technicalReplicatesToRemove = NULL
)
}
\arguments{
\item{rawData}{Data as a tibble}
}
\value{
Lovely filtered data
}
\description{
Filter raw data
}
