% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lipid_class_order}
\alias{get_lipid_class_order}
\title{Lipid Class Order}
\usage{
get_lipid_class_order(con)
}
\arguments{
\item{con}{A database connection
If the databse does not have the table
LIPID_CLASS_ORDER_COMPLETE, the order
is generated here instead.}
}
\value{
A vector of lipid classes
}
\description{
Get or create the order in which the different lipid
classes are displayed on their axis.
}
