% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{imputeMissingIf}
\alias{imputeMissingIf}
\title{Impute implicit missing values as 0}
\usage{
imputeMissingIf(data, doIt = TRUE)
}
\arguments{
\item{data}{tibble. rawData}

\item{doIt}{boolean. Imputes if TRUE,
else returns input data}
}
\value{
tibble. imputed data
}
\description{
Impute implicit missing values as 0
}
