% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.discrete}
\alias{is.discrete}
\title{Test if x is discrete}
\usage{
is.discrete(x)
}
\arguments{
\item{x}{}
}
\value{
boolean
}
\description{
Test if x is discrete
}
\note{
borrowed from plyr (https://github.com/hadley/plyr/)
}
