% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mainScale}
\alias{mainScale}
\title{Create color scale}
\usage{
mainScale(colorCount)
}
\arguments{
\item{colorCount}{integer}
}
\value{
A list with scale_color_ and scale_fill_
}
\description{
Create color scale
}
