% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mainTheme}
\alias{mainTheme}
\title{mainTheme}
\usage{
mainTheme()
}
\value{
list. A list to be added to a ggplot object.
}
\description{
Add \code{+ mainTheme()} to ggplots
to apply the theme used in the shiny app.
}
