% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data_preparation.R
\name{readLipidMapsSDF}
\alias{readLipidMapsSDF}
\title{Read SDF file from Lipid Maps}
\usage{
readLipidMapsSDF(path)
}
\arguments{
\item{path}{string. File path to SDF file}
}
\value{
tibble. Tidy tibble of LM data
}
\description{
Read SDF file from Lipid Maps
}
