% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standardize_rawData}
\alias{standardize_rawData}
\title{Standardize raw data}
\usage{
standardize_rawData(df, base_sample, std_features)
}
\arguments{
\item{df}{Raw data as a tibble}

\item{base_sample}{Sample to use as a baseline}

\item{...}{Features to standardize on}
}
\value{
Standardized data as a tibble
}
\description{
Standardize raw data
}
