% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeRawDataWithin}
\alias{standardizeRawDataWithin}
\title{Standardize raw data}
\usage{
standardizeRawDataWithin(
  data,
  baselineSample = "",
  standardizationFeatures = c("")
)
}
\arguments{
\item{data}{tibble. Raw data}

\item{baselineSample}{string | "", Sample to use as a baseline}

\item{standardizationFeatures}{character vector, Features to standardize on}
}
\value{
tibble. Standardized data
}
\description{
Standardize raw data
}
