% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeWithin}
\alias{standardizeWithin}
\title{Standardize raw data}
\usage{
standardizeWithin(data, input)
}
\arguments{
\item{data}{tibble. Raw data}

\item{input}{list. Input list from shiny ui. Uses
\itemize{
\item baselineSample
\item standardizationFeatures :: tibble. rawData
}}
}
\value{
tibble. Standardized data
}
\description{
Standardize raw data
}
