% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeWithinTechnicalReplicatesIf}
\alias{standardizeWithinTechnicalReplicatesIf}
\title{Standardize data within technical replicates}
\usage{
standardizeWithinTechnicalReplicatesIf(data, doIt)
}
\arguments{
\item{data}{tibble. data}

\item{doIt}{boolean, if TRUE runs standardization,
else returns the data unchanged}
}
\value{
tibble. standardized data
}
\description{
Standardize data within technical replicates
}
