% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stat_chull}
\alias{stat_chull}
\title{Convex hull for PCA plots}
\usage{
stat_chull(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{
a state for ggplot
}
\description{
Borrowed from https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html
}
