% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_plotData}
\alias{summarise_plotData}
\title{Summarise the plot data}
\usage{
summarise_plotData(df)
}
\arguments{
\item{df}{input tibble}
}
\value{
data ready for summmary plots like barplots
}
\description{
Summarise the plot data
}
