% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{summarisePlotData}
\alias{summarisePlotData}
\title{Summarise the plot data}
\usage{
summarisePlotData(
  data,
  aesX = "class",
  aesColor = "sample",
  aesFacetCol = "",
  aesFacetRow = ""
)
}
\arguments{
\item{data}{tibble. plotData}

\item{aesX}{string}

\item{aesColor}{string}

\item{aesFacetCol}{string | ""}

\item{aesFacetRow}{string | ""}
}
\value{
tibble. data ready for summmary plots like barplots
}
\description{
Summarise the plot data
}
