% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{testAllMoreThanOneReplicate}
\alias{testAllMoreThanOneReplicate}
\title{Test that all samples have more than one replicate}
\usage{
testAllMoreThanOneReplicate(data, aesX, aesColor)
}
\arguments{
\item{data}{tibble.}

\item{aesX}{string. Feature on the x-axis.}

\item{aesColor}{string. Feature mapped to the color.}
}
\value{
boolean.
}
\description{
Used before \code{doAllPairwiseComparisons}.
}
