% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{testPairwise}
\alias{testPairwise}
\title{Pairwise t-tests}
\usage{
testPairwise(data)
}
\arguments{
\item{data}{a subset of the data}
}
\value{
A tidy representation of the significance
test build with \code{broom::tidy}
}
\description{
Because people like p-values and little stars.
}
