% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inputs.R
\name{updateAllRangeInputs}
\alias{updateAllRangeInputs}
\title{Custom update range input}
\usage{
updateAllRangeInputs(inputName, choiceColumn, data, session)
}
\arguments{
\item{inputName}{string, input to update}

\item{choiceColumn}{string, column in the data to get choices from}

\item{data}{rawData}

\item{session}{session}
}
\value{
IO ()
}
\description{
Custom update range input
}
