% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inputs.R
\name{updateAllSelectizeInputs}
\alias{updateAllSelectizeInputs}
\title{custom update selectize input}
\usage{
updateAllSelectizeInputs(
  inputName,
  choiceColumn,
  selectedChoice,
  data,
  session
)
}
\arguments{
\item{inputName}{}

\item{choiceColumn}{}

\item{selectedChoice}{}

\item{data}{}

\item{session}{}
}
\value{
IO ()
}
\description{
custom update selectize input
}
