% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inputs.R
\name{updateInputsForSpeciesProfile}
\alias{updateInputsForSpeciesProfile}
\title{Update inputs for species profile}
\usage{
updateInputsForSpeciesProfile(session, quickSpeciesProfileClass)
}
\arguments{
\item{session}{session}

\item{quickSpeciesProfileClass}{string, lipid class to display as profile}
}
\value{
IO ()
}
\description{
Update inputs for species profile
}
