% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collect_raw_data}
\alias{collect_raw_data}
\title{Collect raw data from database}
\usage{
collect_raw_data(con, query, custom_class_order)
}
\arguments{
\item{con}{A database connection}

\item{query}{The sql query to get the dataset.
Create it yourself with \code{\link{sqlQueryData}}.}

\item{custom_class_order}{The order for the lipid classes}
}
\value{
Raw data as a tibble
}
\description{
Collect raw data from database
}
