% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{collectRawData}
\alias{collectRawData}
\title{Collect raw data from database}
\usage{
collectRawData(con, query, lipidClassOrder)
}
\arguments{
\item{con}{A database connection}

\item{query}{The sql query to get the dataset.
Create it yourself with \code{\link{createQueryForID}}.}

\item{lipidClassOrder}{The order for the lipid classes}
}
\value{
Raw data as a tibble
}
\description{
Collect raw data from database
}
