% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{createDatabase}
\alias{createDatabase}
\title{Create your own database}
\usage{
createDatabase(path = "databaseDump.db", meta_info, dataset)
}
\arguments{
\item{path}{string,
Path to where you want to save the database dump file}

\item{meta_info}{data.frame
A data.frame with the information about your datasets
See README on href{https://github.com/jannikbuhr/ShinyLipids}{github} for the columns it needs.}

\item{dataset}{A data.frame with the datasets
See README on href{https://github.com/jannikbuhr/ShinyLipids}{github} for the columns they need.}
}
\description{
Quickly create your own SQLite database dump from
dataframes without additional software.
}
