% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create a Heatmap from data}
\usage{
createHeatmap(data, input)
}
\arguments{
\item{data}{tibble. typically plotData()}

\item{input}{:: list. Uses:
- standardizationFeatures NULL | character vector.
- aesX :: string.
- aesColor :: string.
- aesFacetCol :: string.
- aesFacetRow :: string.
- heatLabSize :: numeric. Size of axis labels
- heatColor :: string. Color palette
  One of c("viridis", "magma", "plasma", "inferno", "cividis")}
}
\value{
ggplot. Heatmap
}
\description{
Create a Heatmap from data
}
