% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create a Heatmap from data}
\usage{
createHeatmap(data, input)
}
\arguments{
\item{data}{a tibble, typically plotData()}

\item{input}{list of inputs from shinys app_ui.R}
}
\value{
a heatmap as a ggplot object
}
\description{
Create a Heatmap from data
}
