% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{createMainPlot}
\alias{createMainPlot}
\title{Title}
\usage{
createMainPlot(
  plotData,
  meanPlotData,
  pairwiseComparisons,
  rangeX,
  rangeY,
  input
)
}
\arguments{
\item{plotData}{data for the plot, pass it from reactive plotData()}

\item{meanPlotData}{data of means, pass from reactive meanPlotData()}

\item{pairwiseComparisons}{a tibble of pairwise t-tests from pairwiseComparisons()}

\item{rangeX}{vector with min and max X}

\item{rangeY}{vector with min and max Y}

\item{input}{list of inputs from shiny UI}
}
\value{
a ggplot object
}
\description{
Title
}
