% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{createMainPlot}
\alias{createMainPlot}
\title{Title}
\usage{
createMainPlot(
  plotData,
  meanPlotData,
  pairwiseComparisons,
  rangeX = NULL,
  rangeY = NULL,
  aesX = "class",
  aesColor = "sample",
  aesFacetCol = "",
  aesFacetRow = "",
  mainPlotAdditionalOptions = list("points", "bars"),
  errorbarType = "None",
  summariseTechnicalReplicates = TRUE,
  standardizationFeatures = NULL,
  standardizeWithinTechnicalReplicate = TRUE
)
}
\arguments{
\item{plotData}{tibble. Data for the plot, pass it from reactive plotData()}

\item{meanPlotData}{tibble. Data of means, pass from reactive meanPlotData()}

\item{pairwiseComparisons}{tibble. Pairwise t-tests from pairwiseComparisons()}

\item{rangeX}{numeric vector | NULL. vector with min and max X}

\item{rangeY}{numeric vector | NULL.  vector with min and max Y}

\item{aesX}{string.}

\item{aesColor}{string.}

\item{aesFacetCol}{string.}

\item{aesFacetRow}{string.}

\item{mainPlotAdditionalOptions}{character vector. Options:
list("points", "bars", "mean", "values", "ind_values", "log", 
"N", "label", "swap", "free_y", "signif")}

\item{errorbarType}{string. "None" | "SD" | "SEM" | "CI"}

\item{summariseTechnicalReplicates}{boolean.}

\item{standardizationFeatures}{character vector | NULL.}

\item{standardizeWithinTechnicalReplicate}{boolean.}
}
\value{
ggplot object
}
\description{
Title
}
