% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{createMainPlot}
\alias{createMainPlot}
\title{Title}
\usage{
createMainPlot(
  plotData,
  meanPlotData,
  pairwiseComparisons,
  rangeX = NULL,
  rangeY = NULL,
  aesX = "class",
  aesColor = "sample",
  aesFacetCol = "",
  aesFacetRow = "",
  mainPlotAdditionalOptions = list("points", "bars"),
  errorbarType = "None",
  summariseTechnicalReplicates = TRUE,
  standardizationFeatures = NULL,
  standardizeWithinTechnicalReplicate = TRUE
)
}
\arguments{
\item{plotData}{data for the plot, pass it from reactive plotData()}

\item{meanPlotData}{data of means, pass from reactive meanPlotData()}

\item{pairwiseComparisons}{a tibble of pairwise t-tests from pairwiseComparisons()}

\item{rangeX}{vector with min and max X}

\item{rangeY}{vector with min and max Y}
}
\value{
a ggplot object
}
\description{
Title
}
