% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{createPlotData}
\alias{createPlotData}
\title{Create the data ready for plotting}
\usage{
createPlotData(data, input)
}
\arguments{
\item{data}{:: tibble}

\item{input}{:: list. Input list from shiny ui. Uses
- summariseTechnicalReplicates :: boolean
- aesX :: string
- aesColor :: string
- aesFacetCol :: string | ""
- aesFacetRow :: string | ""}
}
\value{
:: tibble. neat data
}
\description{
Based on the aesthetic mappings.
}
