% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createPlotData}
\alias{createPlotData}
\title{Create the data ready for plotting}
\usage{
createPlotData(df, input)
}
\arguments{
\item{df}{Input tibble}

\item{input}{a list with the aesthetic mappings.
This list can contain
strings with the following names,
non existent ones will be ignored:
\itemize{
\item \code{summariseTechnicalReplicates}
\item \code{aesX}
\item \code{aesColor}
\item \code{aesFacetCol}
\item \code{aesFacetRow}
}}
}
\value{
neat data
}
\description{
Based on the aesthetic mappings.
}
