% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{createPlotData}
\alias{createPlotData}
\title{Create the data ready for plotting}
\usage{
createPlotData(
  data,
  summariseTechnicalReplicates = TRUE,
  aesX = "class",
  aesColor = "sample",
  aesFacetCol = NULL,
  aesFacetRow = NULL
)
}
\arguments{
\item{data}{Input tibble}
}
\value{
neat data
}
\description{
Based on the aesthetic mappings.
}
