% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{createQueryForID}
\alias{createQueryForID}
\title{Generate SQL Query for the selected dataset}
\usage{
createQueryForID(datasetID)
}
\arguments{
\item{datasetID}{numeric. ID of the dataset}
}
\value{
string. SQL Query
}
\description{
Generate SQL Query for the selected dataset
}
