% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{doAllPairwiseComparisons}
\alias{doAllPairwiseComparisons}
\title{Test all possible pairwise t-tests}
\usage{
doAllPairwiseComparisons(data, input)
}
\arguments{
\item{data}{tibble. The full dataset}

\item{input}{:: list. Input list from shiny ui. Uses
\itemize{
\item aesX :: string
}}
}
\value{
:: tibble.
A tidy representation of the significance
tests build with \code{broom::tidy} combined
into one tibble.
}
\description{
Because people like p-values and little stars.
}
