% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{doAllPairwiseComparisons}
\alias{doAllPairwiseComparisons}
\title{Test all possible pairwise t-tests}
\usage{
doAllPairwiseComparisons(data, aesX)
}
\arguments{
\item{data}{tibble. The full dataset}

\item{aesX}{string. the feature mapped to the x-axis}
}
\value{
tibble.
A tidy representation of the significance
tests build with \code{broom::tidy} combinde
into one tibble.
}
\description{
Because people like p-values and little stars.
}
