% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(df, input)
}
\arguments{
\item{df}{Data as a tibble}

\item{input}{a list.
This list can contain
strings with the following names,
non existent ones will be ignored:
\itemize{
\item \code{categoryToSelect}
\item \code{lipidClassToSelect}
\item \code{functionalCategoryToSelect}
\item \code{filter_length}
\item \code{filter_db}
\item \code{filter_oh}
\item \code{samplesToSelect}
\item \code{samplesToRemove}
\item \code{technicalReplicatesToRemove}
}}
}
\value{
Lovely filtered data
}
\description{
Filter raw data
}
