% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(rawData, input)
}
\arguments{
\item{rawData}{:: tibble. rawData}

\item{input}{:: list. Input list from shiny ui. Uses
- standardizeWithinTechnicalReplicate
- categoryToSelect :: string | NULL
- lipidClassToSelect :: string | NULL
- functionalCategoryToSelect :: string | NULL
- filterLengthRange :: string | NULL
- filterDoubleBondsRange :: string | NULL
- filterOhRange :: string | NULL
- samplesToSelect :: string | NULL
- samplesToRemove :: string | NULL
- replicatesToSelect :: string | NULL
- replicatesToRemove :: string | NULL
- technicalReplicatesToRemove :: string | NULL}
}
\value{
tibble. Lovely filtered data
}
\description{
Filter raw data
}
