% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(
  rawData,
  categoryToSelect = NULL,
  lipidClassToSelect = NULL,
  functionalCategoryToSelect = NULL,
  filterLengthRange = NULL,
  filterDoubleBondsRange = NULL,
  filterOhRange = NULL,
  samplesToSelect = NULL,
  samplesToRemove = NULL,
  replicatesToSelect = NULL,
  replicatesToRemove = NULL,
  technicalReplicatesToRemove = NULL
)
}
\arguments{
\item{rawData}{tibble. rawData}

\item{categoryToSelect}{string | NULL}

\item{lipidClassToSelect}{string | NULL}

\item{functionalCategoryToSelect}{string | NULL}

\item{filterLengthRange}{string | NULL}

\item{filterDoubleBondsRange}{string | NULL}

\item{filterOhRange}{string | NULL}

\item{samplesToSelect}{string | NULL}

\item{samplesToRemove}{string | NULL}

\item{replicatesToSelect}{string | NULL}

\item{replicatesToRemove}{string | NULL}

\item{technicalReplicatesToRemove}{string | NULL}
}
\value{
tibble. Lovely filtered data
}
\description{
Filter raw data
}
