% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filterRawDataFor}
\alias{filterRawDataFor}
\title{Filter raw data}
\usage{
filterRawDataFor(rawData, input)
}
\arguments{
\item{rawData}{:: tibble. rawData}

\item{input}{:: list. Input list from shiny ui. Uses
\itemize{
\item standardizeWithinTechnicalReplicate
\item categoryToSelect :: string | NULL
\item lipidClassToSelect :: string | NULL
\item functionalCategoryToSelect :: string | NULL
\item filterLengthRange :: string | NULL
\item filterDoubleBondsRange :: string | NULL
\item filterOhRange :: string | NULL
\item samplesToSelect :: string | NULL
\item samplesToRemove :: string | NULL
\item replicatesToSelect :: string | NULL
\item replicatesToRemove :: string | NULL
\item technicalReplicatesToRemove :: string | NULL
}}
}
\value{
tibble. Lovely filtered data
}
\description{
Filter raw data
}
