% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{imputeMissingIf}
\alias{imputeMissingIf}
\title{Impute implicit missing values as 0}
\usage{
imputeMissingIf(data, input)
}
\arguments{
\item{data}{tibble. rawData}

\item{input}{list. Input list from shiny ui. Uses
\itemize{
\item imputeMissingAs0 :: logical
}}
}
\value{
tibble. imputed data
}
\description{
Imputes if \code{input$imputeMissingAs0 == TRUE},
else returns input data.
}
