% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standardize_rawData}
\alias{standardize_rawData}
\title{Standardize raw data}
\usage{
standardize_rawData(df, std_feature, base_sample)
}
\arguments{
\item{df}{Raw data as a tibble}

\item{std_feature}{Feature so standardize on}

\item{base_sample}{Sample to use as a baseline}
}
\value{
Standardized data as a tibble
}
\description{
Standardize raw data
}
