% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeRawDataWithin}
\alias{standardizeRawDataWithin}
\title{Standardize raw data}
\usage{
standardizeRawDataWithin(data, input)
}
\arguments{
\item{data}{tibble. Raw data}

\item{input}{list. Input list from shiny ui. Uses
- baselineSample
- standardizationFeatures :: tibble. rawData}
}
\value{
tibble. Standardized data
}
\description{
Standardize raw data
}
