% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeRawDataWithin}
\alias{standardizeRawDataWithin}
\title{Standardize raw data}
\usage{
standardizeRawDataWithin(
  data,
  baselineSample = "",
  standardizationFeatures = c("")
)
}
\arguments{
\item{data}{Raw data as a tibble}

\item{baselineSample}{Sample to use as a baseline}

\item{standardizationFeatures}{Features to standardize on}
}
\value{
Standardized data as a tibble
}
\description{
Standardize raw data
}
