% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{standardizeWithinTechnicalReplicatesIf}
\alias{standardizeWithinTechnicalReplicatesIf}
\title{Standardize data within technical replicates}
\usage{
standardizeWithinTechnicalReplicatesIf(data, input)
}
\arguments{
\item{data}{tibble. data}

\item{input}{list. Input list from shiny ui. Uses
\itemize{
\item standardizeWithinTechnicalReplicate :: logical
}}
}
\value{
tibble. standardized data
}
\description{
Standardize data within technical replicates
}
