% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stat_chull}
\alias{stat_chull}
\title{Convex hull for PCA plots}
\usage{
stat_chull(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data}

\item{geom}{geometric element}

\item{position}{default = "identity"}

\item{na.rm}{remove NAs}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit aesthetics}

\item{...}{passed to layer}
}
\value{
a state for ggplot
}
\description{
Borrowed from https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html
}
