% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{summarisePlotData}
\alias{summarisePlotData}
\title{Summarise the plot data}
\usage{
summarisePlotData(data, input)
}
\arguments{
\item{data}{tibble. plotData}

\item{input}{:: list. Input list from shiny ui. Uses
- aesX :: string
- aesColor :: string
- aesFacetCol :: string | ""
- aesFacetRow :: string | ""}
}
\value{
tibble. data ready for summmary plots like barplots
}
\description{
Summarise the plot data
}
