% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarisePlotData}
\alias{summarisePlotData}
\title{Summarise the plot data}
\usage{
summarisePlotData(df)
}
\arguments{
\item{df}{input tibble}
}
\value{
data ready for summmary plots like barplots
}
\description{
Summarise the plot data
}
