% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_pairwise}
\alias{test_pairwise}
\title{Pairwise t-tests}
\usage{
test_pairwise(df)
}
\arguments{
\item{df}{a subset of the data}
}
\value{
A tidy representation of the significance
test build with \code{\link{broom::tidy}}
}
\description{
Because people like p-values and little stars.
}
